每轮次评价与报告生成

+++ src/main/java/org/buhuiqiming/fuchuang/dto/GenerateReportRequest.java
@@ -60,7 +60,7 @@ public class GenerateReportRequest {
     @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
     public static class DimensionScores {
         private float professional;
-        private Integer cognition;
+        private float cognition;
         private float expression;
     }
